LayoutTests/js/typedarray-set-source-smaller-than-destination-expected.txt

Tests the code path in typedArray.set that may have to do a copy via an intermediate buffer because the source and destination overlap and have different size elements (source is smaller than destination).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS foo(10) is foo_reference(10)
PASS bar(10) is bar_reference(10)
PASS baz(10) is baz_reference(10)
PASS fuz(10) is fuz_reference(10)
PASS thingy(10) is thingy_reference(10)
PASS successfullyParsed is true

TEST COMPLETE
